+++ noodles-vcf/src/header/filter.rs
@@ -313,10 +313,7 @@ mod tests {
 
     #[test]
     fn test_try_from_record_for_filter_with_an_invalid_record_value() {
-        let record = Record::new(
-            record::key::FILTER,
-            record::Value::String(String::from("VCFv4.3")),
-        );
+        let record = Record::new(record::key::FILTER, record::Value::from("VCFv4.3"));
 
         assert_eq!(
             Filter::try_from(record),